requirements.txt
instagrapi>=2.0.0
openai>=1.0.0
openai-whisper>=20231117
python-dotenv>=1.0.0
pillow>=10.0.0
requests>=2.31.0
opencv-python>=4.8.0
moviepy>=1.0.3
tabulate>=0.9.0
